Fix stale price issue. Always supply metadata required by Bisq client.
1. There was an issue that if a provider started rejecting requests,
the pricenode would keep delivering the last known rates to the
Bisq average calculation.  Over time as market moved, the calculated
rate would diverge from reality.  This happened when Poloniex feed
went down and the XMR price became inaccurate.

2. If pricenode was started and a provider was down the metadata
fields would not be supplied to Bisq.  This caused an NPE in Bisq
which is hardcoded to expect certain exchange names in metadata.
Bisq would also reject the pricenodes and as a result have no
market data in that case.  Fix is to always include the metadata
for all known exchanges, even when they are not supplying a price.